requirements/prod.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-decouple==3.8
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.12.2
tzdata==2025.2
